Add entity damage event handling in ProfessionManager
This commit adds a new event handler for EntityDamageByEntityEvent in the ProfessionManager class. The handler checks if the damage is caused by a player and performs the corresponding profession action if applicable. It also sends a message to the player indicating that they have attacked an entity.

The commit also includes some minor changes in the Hunter class, where it increases the damage dealt by the player based on their profession level.

Note: Some code has been commented out for future implementation.